http-proxy-policy/requirements.txt
django==5.1.7
djangorestframework==3.15.2
pydantic==2.10.6
psycopg2-binary==2.9.10